+++ src/CMakeLists.txt
@@ -1,4 +1,6 @@
 target_sources(pairwise
     PRIVATE
-        pairwise.c   pairwise.h  
+        pairwise.c        pairwise.h
+        potentials.c      potentials.h
+        spherocylinder.c  spherocylinder.h
 )
