+++ src/pairwise.h
@@ -1,4 +1,10 @@
-#include <stdio.h>
+/** @file pairwise.h
+ *  @brief Shared definitions for the morpho-pairwise package
+ */
+
+#ifndef PAIRWISE_H
+#define PAIRWISE_H
+
 #include <morpho.h>
 
 #define PAIRWISE_CLASSNAME                   "Pairwise"
@@ -8,14 +14,6 @@
 #define PAIRWISE_PERIODIC_PROPERTY           "box"
 #define PAIRWISE_SIGMA_PROPERTY              "sigma"
 #define PAIRWISE_CENTER_PROPERTY             "center"
-#define LJ_SIGMA_PROPERTY                    "sigma"
-
-#define GRAVITY_CLASSNAME                    "GravityPotential"
-#define COULOMB_CLASSNAME                    "CoulombPotential"
-#define HERTZIAN_CLASSNAME                   "HertzianPotential"
-#define LENNARDJONES_CLASSNAME               "LJPotential"
-
-#define SPHEROCYLINDER_CLASSNAME             "SpherocylinderOverlap"
 
 #define PAIRWISE_VALUE_METHOD                "value"
 #define PAIRWISE_DERIVATIVE_METHOD           "derivative"
@@ -23,12 +21,15 @@
 #define PAIRWISE_PRP                         "PrwsPrp"
 #define PAIRWISE_PRP_MSG                     "Pairwise properties."
 
-#define SPHEROCYLINDER_FLD                   "SphrCylFld"
-#define SPHEROCYLINDER_FLD_MSG               "Spherocylinder requires a Field to define orientation."
-
-#define SPHEROCYLINDER_DIM                   "SphrCylDim"
-#define SPHEROCYLINDER_DIM_MSG               "Dimension of vector field must match dimension of space."
-
-void functional_vecsub_periodic(unsigned int n, double *a, double *b, double box, double *out);
+/* Shared interned symbols used across the package */
+extern value pairwise_potentialproperty;
+extern value pairwise_cutoffproperty;
+extern value pairwise_periodicproperty;
+extern value pairwise_sigmaproperty;
+extern value pairwise_centerproperty;
+extern value pairwise_valuemethod;
+extern value pairwise_derivativemethod;
 
 void pairwise_initialize(void);
+
+#endif
